pyproject.toml
[project]
name = "pushups"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.9"
dependencies = [
    "python-dotenv>=1.1.0",
    "python-telegram-bot[job-queue]>=22.1",
]
